fix(): all comments from PR
