+++ scripts/prebuild.js
@@ -1,8 +1,6 @@
 /* eslint-disable */
 import fs from "fs"
 import dotenv from "dotenv"
-import { fileURLToPath } from "url"
-import { dirname } from "path"
 
 dotenv.config()
