+++ src/App.styled.tsx
@@ -1,0 +1,18 @@
+import { styled } from "styled-components"
+
+const AppContainer = styled.div`
+  background-color: #0f1417;
+  color: #ebecfa;
+  position: relative;
+  width: 100%;
+  height: 100%;
+  min-height: calc(100vh - 50px);
+  background-repeat: no-repeat;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: flex-start;
+  position: relative;
+`
+
+export { AppContainer }
